+++ lib/water_drop/async_producer.rb
@@ -1,0 +1,5 @@
+module WaterDrop
+  # Async producer for messages
+  AsyncProducer = Class.new(BaseProducer)
+  AsyncProducer.method_name = :deliver_async
+end