+++ lib/water_drop/sync_producer.rb
@@ -1,0 +1,5 @@
+module WaterDrop
+  # Sync producer for messages
+  Producer = SyncProducer = Class.new(BaseProducer)
+  SyncProducer.method_name = :deliver
+end