+++ spec/lib/water_drop_spec.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'logger'
-
 RSpec.describe WaterDrop do
   describe '#logger' do
     let(:logger) { double }